[Added] ABC 032A, 041C, 058B, 086C, 155C, 187D, ARC 089A

+++ AtCoder/Libraries/数字系/切り上げ除法.cpp
@@ -3,7 +3,7 @@ using namespace std;
 
 typedef long long ll;
 
-int RoundUpDivision(ll a, ll b) {
+ll RoundUpDivision(ll a, ll b) {
     ll ans = (a + b - 1) / b;
     return ans;
 }
